Add dimension assertion and clarify accuracy/Hessian calculations
Agent-Logs-Url: https://github.com/OreanHeinrich/CA2_SVM/sessions/d9cded93-cebb-4e7d-871d-2f9672ccf417

+++ task2_evaluate_table1.m
@@ -59,6 +59,6 @@ if isempty(model)
 end
 predTr = svm_predict_labels(model, Xtr);
 predTe = svm_predict_labels(model, Xte);
-accTr = mean(predTr == ytr);
-accTe = mean(predTe == yte);
+accTr = mean(double(predTr == ytr));
+accTe = mean(double(predTe == yte));
 end
